Fix Poly1305 bug in fast version and bring it back
See

https://github.com/floodyberry/poly1305-donna/commit/0911057b9607f37e642337739a2834aec84fbb53

for fix in original C version.

"The issue was not properly detecting if st->h was >= 2^130 - 5, coupled
with believing my own README and defining POLY1305_16BITS instead of
POLY1305_16BIT and not catching the failure. The chance of the bug
affecting anything in the real world is essentially zero luckily, but
it's good to have it fixed."

Thanks to @floodyberry for fixing it quickly.