Added the name of the license
I created [legally](http://github.com/franciscop/legally) and this is one of the few packages that stands out for not having the short name for the license int he package. I added it. It's also not normal to put it in `COPYING.txt`, normally it goes into `LICENSE.txt`, but that is handled with my library.